feat: branding.translationProvider + tafsirProvider seams (RFC-009)
Implements the RFC-009 content-provider seams. Adds two optional
`Branding` slots — `translationProvider` and `tafsirProvider` — so
forks can supply their own translation / tafsir sources for the
Settings → Translations and Settings → Tafsir download flows.
Bayaan's behaviour is unchanged when the slots are unset.

- New `types/TranslationProvider.ts` + `types/TafsirProvider.ts`
  interfaces (two methods each, mirroring the existing service
  surfaces).
- `TranslationVerse` / `TafseerVerse` moved to `types/translation.ts`
  / `types/tafseer.ts` so the interfaces + implementations can share
  them; the old locations re-export for back-compat.
- The current service class bodies are extracted verbatim into
  default-named providers: `AlQuranCloudTranslationProvider` and
  `QuranComTafsirProvider`.
- `TranslationApiService.ts` / `TafseerApiService.ts` become thin
  resolvers — `branding.<provider> ?? <default>`. The exported
  singleton names + shapes are unchanged, so `store/translationStore`
  and `store/tafseerStore` are untouched.

No behaviour change with both slots unset: base URLs, response
shapes, the RTL `direction` fallback, the paginated by-chapter loop,
and the verse-grouping logic are all preserved verbatim.

RFC: docs/rfcs/009-content-provider-seams.md (merged via #262).
tsc --noEmit clean.
